pyproject.toml
[tool.poetry]
name = "Python WEB HW 13 (Fast API part)"
version = "0.1.0"
description = "A FastAPI project to save contacts. Authorization added."
authors = ["Yuliia Melnychenko <arwen.vogel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11.5" 
fastapi = "0.115.0"
fastapi_mail = "1.4.1"
sqlalchemy = "2.0.35"
psycopg2 = "2.9.9"
alembic = "1.13.3"
uvicorn = { version = "0.31.1", extras = ["standard"] }
pydantic = { version = "2.9.2", extras = ["email"] }
email_validator = "2.2.0"
libgravatar = "^1.0.4"
python-jose = "^3.3.0"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-multipart = "^0.0.12"
redis = "^5.2.0"
pydantic-settings = "^2.6.0"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
